[BE-154] Replace fns get_exchange_rate/calculate_exchange_rate/get_wrap_by_underlying/get_underlying_by_wrap from controller to pool. Clean up Errors in controller pallet.
