+++ pallets/controller/src/lib.rs
@@ -55,9 +55,6 @@ pub trait Trait: liquidity_pools::Trait + system::Trait + oracle::Trait + accoun
 	/// The overarching event type.
 	type Event: From<Event> + Into<<Self as system::Trait>::Event>;
 
-	/// Start exchange rate
-	type InitialExchangeRate: Get<Rate>;
-
 	/// The approximate number of blocks per year
 	type BlocksPerYear: Get<u128>;
 
@@ -113,12 +110,6 @@ decl_error! {
 
 		/// Borrow rate is absurdly high.
 		BorrowRateIsTooHight,
-
-		/// The currency is not enabled in protocol.
-		NotValidUnderlyingAssetId,
-
-		/// The currency is not enabled in wrapped protocol.
-		NotValidWrappedTokenId,
 
 		/// Oracle unavailable or price equal 0ю
 		OraclePriceError,
@@ -338,7 +329,6 @@ decl_module! {
 type RateResult = result::Result<Rate, DispatchError>;
 type BalanceResult = result::Result<Balance, DispatchError>;
 type LiquidityResult = result::Result<(Balance, Balance), DispatchError>;
-type CurrencyIdResult = result::Result<CurrencyId, DispatchError>;
 
 impl<T: Trait> Module<T> {
 	/// Applies accrued interest to total borrows and insurances.
@@ -689,52 +679,52 @@ impl<T: Trait> Module<T> {
 
 // Private methods
 impl<T: Trait> Module<T> {
-	/// Calculates the exchange rate from the underlying to the mToken.
-	/// This function does not accrue interest before calculating the exchange rate.
-	pub fn get_exchange_rate(underlying_asset_id: CurrencyId) -> RateResult {
-		let wrapped_asset_id = Self::get_wrapped_id_by_underlying_asset_id(&underlying_asset_id)?;
-		// The total amount of cash the market has
-		let total_cash = <LiquidityPools<T>>::get_pool_available_liquidity(underlying_asset_id);
-
-		// Total number of tokens in circulation
-		let total_supply = T::MultiCurrency::total_issuance(wrapped_asset_id);
-
-		let total_insurance = <LiquidityPools<T>>::get_pool_total_insurance(underlying_asset_id);
-
-		let total_borrowed = <LiquidityPools<T>>::get_pool_total_borrowed(underlying_asset_id);
-
-		let current_exchange_rate =
-			Self::calculate_exchange_rate(total_cash, total_supply, total_insurance, total_borrowed)?;
-		// FIXME: can be removed.
-		<LiquidityPools<T>>::set_current_exchange_rate(underlying_asset_id, current_exchange_rate)?;
-
-		Ok(current_exchange_rate)
-	}
-
-	/// Calculates the exchange rate from the underlying to the mToken.
-	fn calculate_exchange_rate(
-		total_cash: Balance,
-		total_supply: Balance,
-		total_insurance: Balance,
-		total_borrowed: Balance,
-	) -> RateResult {
-		let rate = match total_supply.cmp(&Balance::zero()) {
-			// If there are no tokens minted: exchangeRate = InitialExchangeRate.
-			Ordering::Equal => T::InitialExchangeRate::get(),
-			// Otherwise: exchange_rate = (total_cash - total_insurance + total_borrowed) / total_supply
-			_ => {
-				let cash_plus_borrows = total_cash.checked_add(total_borrowed).ok_or(Error::<T>::NumOverflow)?;
-
-				let cash_plus_borrows_minus_insurance = cash_plus_borrows
-					.checked_sub(total_insurance)
-					.ok_or(Error::<T>::NumOverflow)?;
-
-				Rate::saturating_from_rational(cash_plus_borrows_minus_insurance, total_supply)
-			}
-		};
-
-		Ok(rate)
-	}
+	// /// Calculates the exchange rate from the underlying to the mToken.
+	// /// This function does not accrue interest before calculating the exchange rate.
+	// pub fn get_exchange_rate(underlying_asset_id: CurrencyId) -> RateResult {
+	// 	let wrapped_asset_id = Self::get_wrapped_id_by_underlying_asset_id(&underlying_asset_id)?;
+	// 	// The total amount of cash the market has
+	// 	let total_cash = <LiquidityPools<T>>::get_pool_available_liquidity(underlying_asset_id);
+	//
+	// 	// Total number of tokens in circulation
+	// 	let total_supply = T::MultiCurrency::total_issuance(wrapped_asset_id);
+	//
+	// 	let total_insurance = <LiquidityPools<T>>::get_pool_total_insurance(underlying_asset_id);
+	//
+	// 	let total_borrowed = <LiquidityPools<T>>::get_pool_total_borrowed(underlying_asset_id);
+	//
+	// 	let current_exchange_rate =
+	// 		Self::calculate_exchange_rate(total_cash, total_supply, total_insurance, total_borrowed)?;
+	// 	// FIXME: can be removed.
+	// 	<LiquidityPools<T>>::set_current_exchange_rate(underlying_asset_id, current_exchange_rate)?;
+	//
+	// 	Ok(current_exchange_rate)
+	// }
+	//
+	// /// Calculates the exchange rate from the underlying to the mToken.
+	// fn calculate_exchange_rate(
+	// 	total_cash: Balance,
+	// 	total_supply: Balance,
+	// 	total_insurance: Balance,
+	// 	total_borrowed: Balance,
+	// ) -> RateResult {
+	// 	let rate = match total_supply.cmp(&Balance::zero()) {
+	// 		// If there are no tokens minted: exchangeRate = InitialExchangeRate.
+	// 		Ordering::Equal => T::InitialExchangeRate::get(),
+	// 		// Otherwise: exchange_rate = (total_cash - total_insurance + total_borrowed) / total_supply
+	// 		_ => {
+	// 			let cash_plus_borrows = total_cash.checked_add(total_borrowed).ok_or(Error::<T>::NumOverflow)?;
+	//
+	// 			let cash_plus_borrows_minus_insurance = cash_plus_borrows
+	// 				.checked_sub(total_insurance)
+	// 				.ok_or(Error::<T>::NumOverflow)?;
+	//
+	// 			Rate::saturating_from_rational(cash_plus_borrows_minus_insurance, total_supply)
+	// 		}
+	// 	};
+	//
+	// 	Ok(rate)
+	// }
 
 	/// Calculates the current borrow rate per block.
 	fn calculate_borrow_interest_rate(underlying_asset_id: CurrencyId, utilization_rate: Rate) -> RateResult {
@@ -900,25 +890,25 @@ impl<T: Trait> Module<T> {
 		Ok(result)
 	}
 
-	pub fn get_wrapped_id_by_underlying_asset_id(asset_id: &CurrencyId) -> CurrencyIdResult {
-		match asset_id {
-			CurrencyId::DOT => Ok(CurrencyId::MDOT),
-			CurrencyId::KSM => Ok(CurrencyId::MKSM),
-			CurrencyId::BTC => Ok(CurrencyId::MBTC),
-			CurrencyId::ETH => Ok(CurrencyId::METH),
-			_ => Err(Error::<T>::NotValidUnderlyingAssetId.into()),
-		}
-	}
-
-	pub fn get_underlying_asset_id_by_wrapped_id(wrapped_id: &CurrencyId) -> CurrencyIdResult {
-		match wrapped_id {
-			CurrencyId::MDOT => Ok(CurrencyId::DOT),
-			CurrencyId::MKSM => Ok(CurrencyId::KSM),
-			CurrencyId::MBTC => Ok(CurrencyId::BTC),
-			CurrencyId::METH => Ok(CurrencyId::ETH),
-			_ => Err(Error::<T>::NotValidWrappedTokenId.into()),
-		}
-	}
+	// pub fn get_wrapped_id_by_underlying_asset_id(asset_id: &CurrencyId) -> CurrencyIdResult {
+	// 	match asset_id {
+	// 		CurrencyId::DOT => Ok(CurrencyId::MDOT),
+	// 		CurrencyId::KSM => Ok(CurrencyId::MKSM),
+	// 		CurrencyId::BTC => Ok(CurrencyId::MBTC),
+	// 		CurrencyId::ETH => Ok(CurrencyId::METH),
+	// 		_ => Err(Error::<T>::NotValidUnderlyingAssetId.into()),
+	// 	}
+	// }
+	//
+	// pub fn get_underlying_asset_id_by_wrapped_id(wrapped_id: &CurrencyId) -> CurrencyIdResult {
+	// 	match wrapped_id {
+	// 		CurrencyId::MDOT => Ok(CurrencyId::DOT),
+	// 		CurrencyId::MKSM => Ok(CurrencyId::KSM),
+	// 		CurrencyId::MBTC => Ok(CurrencyId::BTC),
+	// 		CurrencyId::METH => Ok(CurrencyId::ETH),
+	// 		_ => Err(Error::<T>::NotValidWrappedTokenId.into()),
+	// 	}
+	// }
 }
 
 // Getters for Controller Data
